legal_app/backend/requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
supabase==0.5.8
redis==4.3.4
gunicorn==20.1.0
python-dotenv==0.19.2
requests==2.28.2
markupsafe==2.0.1
itsdangerous==2.0.1
jinja2==3.0.3
importlib-metadata==4.13.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5